fix : AlarmBox Patch

+++ market/src/main/java/com/charly/market/alarmbox/repository/AlarmBoxRepository.java
@@ -2,7 +2,7 @@ package com.charly.market.alarmbox.repository;
 
 import com.charly.market.alarmbox.model.entity.AlarmBox;
 import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.Optional;
 
-public interface AlarmBoxRepository extends JpaRepository<AlarmBox,Integer> {
-
+public interface AlarmBoxRepository extends JpaRepository<AlarmBox, Long> {
 }
